requirements.txt
numpy
pandas
kagglehub
tensorflow
scikit-learn
opencv-python
matplotlib
seaborn
Pillow
